use GTControlledRegistry
