+++ src/main/java/gregtech/common/covers/CoverBehaviors.java
@@ -21,9 +21,12 @@ import net.minecraft.util.EnumFacing;
 import net.minecraft.util.ResourceLocation;
 
 import javax.annotation.Nonnull;
+import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.BiFunction;
 
 public class CoverBehaviors {
+
+    private static final AtomicInteger atomicId = new AtomicInteger();
 
     public static void init() {
         GTLog.logger.info("Registering cover behaviors...");
@@ -169,7 +172,7 @@ public class CoverBehaviors {
     public static CoverDefinition registerCover(@Nonnull ResourceLocation coverId, @Nonnull MetaValueItem itemStack,
                                                 @Nonnull BiFunction<ICoverable, EnumFacing, CoverBehavior> behaviorCreator) {
         CoverDefinition coverDefinition = new CoverDefinition(coverId, behaviorCreator, itemStack.getStackForm());
-        GregTechAPI.COVER_REGISTRY.putObject(coverId, coverDefinition);
+        GregTechAPI.COVER_REGISTRY.register(atomicId.getAndIncrement(), coverId, coverDefinition);
         return coverDefinition;
     }
 }
